Move InvalidHttpResponseException to transport package as that is where it is used from
