+++ client/src/main/java/org/realityforge/replicant/client/gwt/GwtWebPollerTransport.java
@@ -11,6 +11,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.gwt.webpoller.client.AbstractHttpRequestFactory;
 import org.realityforge.gwt.webpoller.client.RequestFactory;
+import org.realityforge.replicant.client.transport.InvalidHttpResponseException;
 import org.realityforge.replicant.client.transport.WebPollerConfig;
 import org.realityforge.replicant.client.transport.WebPollerTransport;
 import org.realityforge.replicant.shared.SharedConstants;
